refactor(packages): apply package by layer to each feature
