+++ src/main/java/nextstep/subway/common/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.shared.exception;
+package nextstep.subway.common.exception;
 
 public class NotFoundException extends RuntimeException {
     public NotFoundException(String message) {
