+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line;
+package nextstep.subway.line.domain;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
@@ -8,7 +8,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import nextstep.subway.station.Station;
+import nextstep.subway.station.domain.Station;
 
 @Entity
 public class Line {
