+++ src/main/java/nextstep/subway/line/domain/LineName.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line;
+package nextstep.subway.line.domain;
 
 import java.util.Objects;
 import java.util.Optional;
