+++ src/test/java/nextstep/subway/line/domain/LineColorTest.java
@@ -1,9 +1,9 @@
-package nextstep.subway.line;
+package nextstep.subway.line.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import nextstep.subway.line.LineColor;
+import nextstep.subway.line.domain.LineColor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
