+++ src/test/java/nextstep/subway/line/domain/LineDistanceTest.java
@@ -1,9 +1,9 @@
-package nextstep.subway.line;
+package nextstep.subway.line.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import nextstep.subway.line.LineDistance;
+import nextstep.subway.line.domain.LineDistance;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
