+++ src/test/java/nextstep/subway/line/domain/LineNameTest.java
@@ -1,9 +1,9 @@
-package nextstep.subway.line;
+package nextstep.subway.line.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import nextstep.subway.line.LineName;
+import nextstep.subway.line.domain.LineName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
